feat(editor,graph): split oversized files and fix all lint errors
- Extract EditorToolbar from Editor.tsx (591→~475 LOC)
- Extract SaveSnapshotModal, SnapshotBrowserModal, graph-schemas from GraphControls.tsx (561→~250 LOC)
- Extract GraphSyncEvents hook from GraphView.tsx (504→~450 LOC)
- Add coverage badge to README.md
- Fix all pre-existing lint errors (81→0)
- Fix any types in sync adapters
- Fix floating promises in MindMapView
- Fix unused imports and variables
- All tests pass (371/371)
- Typecheck passes
- Build passes
