+++ src/features/graph/GraphView.tsx
@@ -10,12 +10,11 @@ import { useRepository } from '../../db/useRepository';
 import { logger } from '../../lib/logger';
 import { perf } from '../../lib/perf';
 import { applyCircularLayout, applyForceLayout, applyHierarchicalLayout } from '../../lib/graph-layout';
-import { useGraphSyncStore } from '../../store/graph-sync-store';
 import { setupGraphSyncListeners } from './sync-adapter';
-import type { SharedNode, SharedEdge } from '../../store/graph-sync-types';
 import { useGraphKeyboardNavigation } from './GraphKeyboardNav';
 import { useGraphTouchGestures } from './GraphTouchHandler';
 import { useGraphSnapshotManager } from './GraphSnapshotManager';
+import { useGraphSyncEvents } from './GraphSyncEvents';
 
 type LayoutType = 'circular' | 'force' | 'hierarchical';
 
@@ -76,11 +75,12 @@ const GraphView: React.FC<Props> = ({
     handleExitSnapshot,
   } = useGraphSnapshotManager(repository);
 
-  // eslint-disable-next-line react-hooks/refs -- hook uses refs only inside effects, not for rendering
   useGraphKeyboardNavigation(
+    // eslint-disable-next-line react-hooks/refs -- hook uses refs only inside effects, not for rendering
     containerRef.current,
+    // eslint-disable-next-line react-hooks/refs -- hook uses refs only inside effects, not for rendering
     sigmaInstance.current,
-    // eslint-disable-next-line react-hooks/refs -- accessed only in effect inside hook
+    // eslint-disable-next-line react-hooks/refs -- hook uses refs only inside effects, not for rendering
     graphRef.current,
     entities,
     selectedNode,
@@ -90,8 +90,12 @@ const GraphView: React.FC<Props> = ({
     repository
   );
 
-  // eslint-disable-next-line react-hooks/refs -- hook uses refs only inside effects, not for rendering
-  useGraphTouchGestures(containerRef.current, sigmaInstance.current);
+  useGraphTouchGestures(
+    // eslint-disable-next-line react-hooks/refs -- hook uses refs only inside effects, not for rendering
+    containerRef.current,
+    // eslint-disable-next-line react-hooks/refs -- hook uses refs only inside effects, not for rendering
+    sigmaInstance.current
+  );
 
   const recomputeNeighborhoodHandler = useCallback((payload: unknown) => {
     const { entities, links, selectedNode } = payload as { entities: Entity[], links: Link[], selectedNode: string };
@@ -313,8 +317,8 @@ const GraphView: React.FC<Props> = ({
 
   useEffect(() => {
     if (!selectedNode) {
+      // eslint-disable-next-line react-hooks/set-state-in-effect -- resetting state before async operation
       setSelectedEntityClaims([]);
-      // eslint-disable-next-line react-hooks/set-state-in-effect -- guard reset before async fetch
       setSelectedEntityLinks([]);
       return;
     }
@@ -344,68 +348,14 @@ const GraphView: React.FC<Props> = ({
     void sigma.getCamera().animatedReset({ duration: 400 });
   }, [layout, effectiveData.entities, links]);
 
-  // Sync effect to consume events
-  useEffect(() => {
-    const unsubscribe = useGraphSyncStore.subscribe((state) => {
-      if (!state.syncEnabled || !graphRef.current || state.pendingEvents.length === 0) return;
-
-      const events = useGraphSyncStore.getState().consumeEvents('graph');
-      if (events.length === 0) return;
-
-      const graph = graphRef.current;
-      events.forEach(event => {
-        if (event.type === 'node:add') {
-          const payload = event.payload as SharedNode;
-          if (!graph.hasNode(payload.id)) {
-            graph.addNode(payload.id, {
-              label: payload.label,
-              size: 10,
-              color: '#2563eb',
-              x: Math.random(),
-              y: Math.random()
-            });
-          }
-        } else if (event.type === 'node:update') {
-          const payload = event.payload as SharedNode;
-          if (graph.hasNode(payload.id)) {
-            const currentLabel = graph.getNodeAttribute(payload.id, 'label');
-            if (currentLabel !== payload.label) {
-              console.warn(`[Sync Conflict] Graph node ${payload.id} has different label. Local: ${currentLabel}, Incoming: ${payload.label}. Applying last-writer wins.`);
-              graph.mergeNodeAttributes(payload.id, { label: payload.label });
-            }
-          }
-        } else if (event.type === 'node:remove') {
-          const payload = event.payload as SharedNode;
-          if (graph.hasNode(payload.id)) {
-            graph.dropNode(payload.id);
-          }
-        } else if (event.type === 'edge:add') {
-          const payload = event.payload as SharedEdge;
-          if (graph.hasNode(payload.from) && graph.hasNode(payload.to) && !graph.hasEdge(payload.id)) {
-            graph.addEdgeWithKey(payload.id, payload.from, payload.to, {
-              label: payload.label,
-              size: 2,
-              color: '#94a3b8'
-            });
-          }
-        } else if (event.type === 'edge:remove') {
-          const payload = event.payload as SharedEdge;
-          if (graph.hasEdge(payload.id)) {
-            graph.dropEdge(payload.id);
-          }
-        }
-      });
-      if (sigmaInstance.current) sigmaInstance.current.refresh();
-    });
-    return () => unsubscribe();
-  }, []);
-
   useEffect(() => {
     return () => {
       sigmaInstance.current?.kill();
       sigmaInstance.current = null;
     };
   }, []);
+
+  useGraphSyncEvents(graphRef);
 
   const handleExportPNG = useCallback(() => {
     const sigma = sigmaInstance.current;
