+++ src/features/graph/graph-schemas.ts
@@ -1,0 +1,16 @@
+import { z } from 'zod';
+
+export const GraphNodeSchema = z.object({
+  id: z.string(),
+  label: z.string(),
+});
+
+export const GraphEdgeSchema = z.object({
+  id: z.string(),
+  source: z.string(),
+  target: z.string(),
+  label: z.string().optional(),
+});
+
+export type GraphNode = z.infer<typeof GraphNodeSchema>;
+export type GraphEdge = z.infer<typeof GraphEdgeSchema>;
