+++ src/features/graph/sync-adapter.ts
@@ -1,55 +1,61 @@
 import type Graph from 'graphology';
 import { useGraphSyncStore } from '../../store/graph-sync-store';
-import { SharedNode, SharedEdge } from '../../store/graph-sync-types';
+
+interface GraphEvent {
+  key: string;
+  attributes?: { label?: string };
+  source?: string;
+  target?: string;
+}
 
 export function setupGraphSyncListeners(
   graph: Graph
 ) {
   const store = useGraphSyncStore.getState();
 
-  const onNodeAdded = ({ key, attributes }: any) => {
+  const onNodeAdded = ({ key, attributes }: GraphEvent) => {
     if (!useGraphSyncStore.getState().syncEnabled) return;
     store.emitEvent({
       type: 'node:add',
       source: 'graph',
-      payload: { id: key, label: attributes.label } as SharedNode,
+      payload: { id: key, label: attributes?.label ?? '' },
     });
   };
 
-  const onNodeAttributesUpdated = ({ key, attributes }: any) => {
+  const onNodeAttributesUpdated = ({ key, attributes }: GraphEvent) => {
     if (!useGraphSyncStore.getState().syncEnabled) return;
     store.emitEvent({
       type: 'node:update',
       source: 'graph',
-      payload: { id: key, label: attributes.label } as SharedNode,
+      payload: { id: key, label: attributes?.label ?? '' },
     });
   };
 
-  const onNodeDropped = ({ key }: any) => {
+  const onNodeDropped = ({ key }: GraphEvent) => {
     if (!useGraphSyncStore.getState().syncEnabled) return;
     store.emitEvent({
       type: 'node:remove',
       source: 'graph',
-      payload: { id: key, label: '' } as SharedNode,
+      payload: { id: key, label: '' },
     });
   };
 
-  const onEdgeAdded = ({ key, source, target, attributes }: any) => {
+  const onEdgeAdded = ({ key, source, target, attributes }: GraphEvent) => {
     if (!useGraphSyncStore.getState().syncEnabled) return;
     store.emitEvent({
       type: 'edge:add',
       source: 'graph',
-      payload: { id: key, from: source, to: target, label: attributes.label } as SharedEdge,
+      payload: { id: key, from: source ?? '', to: target ?? '', label: attributes?.label ?? '' },
     });
   };
 
-  const onEdgeDropped = ({ key }: any) => {
+  const onEdgeDropped = ({ key }: GraphEvent) => {
     if (!useGraphSyncStore.getState().syncEnabled) return;
     // We don't have from/to here, but id should be enough for removal if shared
     store.emitEvent({
       type: 'edge:remove',
       source: 'graph',
-      payload: { id: key, from: '', to: '' } as SharedEdge,
+      payload: { id: key, from: '', to: '' },
     });
   };
