+++ src/features/mindmap/sync-adapter.ts
@@ -1,31 +1,40 @@
 import type { MindElixirInstance } from 'mind-elixir';
 import { useGraphSyncStore } from '../../store/graph-sync-store';
-import { SharedNode } from '../../store/graph-sync-types';
+
+interface MindMapOperation {
+  name: string;
+  obj?: { id: string; topic: string };
+  objs?: Array<{ id: string; topic: string }>;
+}
 
 export function setupMindMapSyncListeners(
   mindMap: MindElixirInstance
 ) {
   const store = useGraphSyncStore.getState();
 
-  mindMap.bus.addListener('operation', (op: any) => {
+  mindMap.bus.addListener('operation', (op: MindMapOperation) => {
     if (!useGraphSyncStore.getState().syncEnabled) return;
 
     if (op.name === 'addChild' || op.name === 'insertSibling') {
       const obj = op.obj;
-      store.emitEvent({
-        type: 'node:add',
-        source: 'mindmap',
-        payload: { id: obj.id, label: obj.topic } as SharedNode,
-      });
+      if (obj) {
+        store.emitEvent({
+          type: 'node:add',
+          source: 'mindmap',
+          payload: { id: obj.id, label: obj.topic },
+        });
+      }
     }
 
     if (op.name === 'finishEdit') {
       const obj = op.obj;
-      store.emitEvent({
-        type: 'node:update',
-        source: 'mindmap',
-        payload: { id: obj.id, label: obj.topic } as SharedNode,
-      });
+      if (obj) {
+        store.emitEvent({
+          type: 'node:update',
+          source: 'mindmap',
+          payload: { id: obj.id, label: obj.topic },
+        });
+      }
     }
 
     if (op.name === 'removeNodes') {
@@ -35,7 +44,7 @@ export function setupMindMapSyncListeners(
           store.emitEvent({
             type: 'node:remove',
             source: 'mindmap',
-            payload: { id: obj.id, label: '' } as SharedNode,
+            payload: { id: obj.id, label: '' },
           });
         });
       }
